+++ grapherrql-package/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 import Header from './components/Header';
-import { GraphContextProvider } from './components/Context';
+import { GraphContextProvider } from './components/SandboxMode/Context';
 
 function App() {
   return (
