+++ grapherrql-package/src/components/SandboxMode/ResponseBox.jsx
